rag-chatbot/requirements.txt
# RAG Chatbot Dependencies
# Backend
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# OpenAI
openai>=1.3.5

# Qdrant Vector Database
qdrant-client>=1.7.0

# Database (optional, for feedback storage)
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0

# Development
httpx>=0.25.2

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
